refactored code and added search by owner
